djangoRest/requirements.txt
Django==1.11.3
djangorestframework==3.6.3
django-cors-headers==2.1.0
pytz==2017.2
django-rest-swagger==2.1.2
numpy==1.13.3

